add ha transfer protocol

+++ store/src/main/java/org/apache/rocketmq/store/ha/autoswitch/AutoSwitchHAService.java
@@ -75,7 +75,7 @@ public class AutoSwitchHAService extends DefaultHAService {
     }
 
     @Override public boolean changeToMaster(int masterEpoch) {
-        final int lastEpoch = this.epochCache.lastEpoch();
+        final long lastEpoch = this.epochCache.lastEpoch();
         if (masterEpoch < lastEpoch) {
             return false;
         }
@@ -105,7 +105,7 @@ public class AutoSwitchHAService extends DefaultHAService {
     }
 
     @Override public boolean changeToSlave(String newMasterAddr, int newMasterEpoch, Long slaveId) {
-        final int lastEpoch = this.epochCache.lastEpoch();
+        final long lastEpoch = this.epochCache.lastEpoch();
         if (newMasterEpoch <= lastEpoch) {
             return false;
         }
